requirements.txt
cffi==1.5.2
cryptography==1.3
enum34==1.1.2
github3.py==1.0.0a4
idna==2.1
ipaddress==1.0.16
ndg-httpsclient==0.4.0
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
requests==2.9.1
six==1.10.0
uritemplate.py==0.3.0
